Frame Extraction 예외 추가 및 로그 도입
