+++ remora/src/main/java/remora/remora/FrameExtraction/FrameExtractionController.java
@@ -1,5 +1,7 @@
 package remora.remora.FrameExtraction;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -14,6 +16,7 @@ import javax.validation.Valid;
 @RestController
 public class FrameExtractionController {
     private final FrameExtractionService frameExtractionService;
+    private Logger log = LoggerFactory.getLogger(getClass());
 
     @Autowired
     public FrameExtractionController(FrameExtractionService frameExtractionService) {
@@ -29,13 +32,14 @@ public class FrameExtractionController {
             for (String path : request.videoCode) {
                 response.frameSet.add(frameExtractionService.frameExtract(path));
             }
-
+            log.info("Frame extraction is {}", "success");
             response.success = true;
-            response.message = "success";
+            response.message = "Success";
             response.videoCode = request.videoCode;
+
         } catch (Exception e) {
             e.printStackTrace();
-
+            log.info("Frame extraction is {}", "fail");
             response.success = false;
             response.message = e.toString();
             response.frameSet = null;
